feat: linked tons of old keryboard stuffs and other

+++ src/isr80h/heap.c
@@ -5,7 +5,7 @@
 
 void *isr80h_command4_malloc(struct interrupt_frame *frame)
 {
-    size_t size = (int)task_get_stack_item(task_current(), 0);
+    size_t size = (uintptr_t)task_get_stack_item(task_current(), 0);
     return process_malloc(task_current()->process, size);
 }
