+++ src/isr80h/io.c
@@ -18,12 +18,12 @@ void *isr80h_command1_print(struct interrupt_frame *frame)
 void *isr80h_command2_getkey(struct interrupt_frame *frame)
 {
     char c = keyboard_pop();
-    return (void *)((int)c);
+    return (void *)((uintptr_t)c);
 }
 
 void *isr80h_command3_putchar(struct interrupt_frame *frame)
 {
-    char c = (char)(int)task_get_stack_item(task_current(), 0);
+    char c = (char)(uintptr_t)task_get_stack_item(task_current(), 0);
     terminal_writechar(c, 15);
     return 0;
 }
