+++ src/isr80h/misc.c
@@ -4,8 +4,8 @@
 
 void *isr80h_command0_sum(struct interrupt_frame *frame)
 {
-    int v2 = (int)task_get_stack_item(task_current(), 1);
-    int v1 = (int)task_get_stack_item(task_current(), 0);
+    intptr_t v2 = (intptr_t)task_get_stack_item(task_current(), 1);
+    intptr_t v1 = (intptr_t)task_get_stack_item(task_current(), 0);
 
     return (void *)(v1 + v2);
 }
